requirements.txt
docx==0.2.4
fitz==0.0.1.dev2
beautifulsoup4==4.10.0
pytesseract==0.3.8
pylatexenc==2.10
tifffile==2021.8.8
Flask==2.1.0
werkzeug==2.0.3
huggingface_hub
chromadb
langchain
pypdf2
sentence-transformers
langchain
langchain_community
skillNer
spacy
en_core_web_lg
fuzzywuzzy
gensim
skillNer==1.0.3
pandas==1.5.3
nltk==3.8.1
spacy==3.7.4
jellyfish==1.0.3
click==8.1.7
joblib==1.3.2
regex==2023.12.25
tqdm==4.66.2
python-dateutil==2.8.2
pytz==2023.4
fuzzywuzzy==0.18.0
pandas==1.5.3
PyMuPDF
python-docx
pytesseract
beautifulsoup4
tifffile
Pillow
